requirements.txt
# ELA Tamper Detection — Dependencies
# Install with: pip install -r requirements.txt

# ─── Deep Learning ──────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0

# ─── Machine Learning ───────────────────────────────────────
scikit-learn>=1.3.0

# ─── Image Processing ───────────────────────────────────────
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# ─── Web Framework ──────────────────────────────────────────
Django>=4.2.0
Pillow>=10.0.0   # needed for ImageField

# ─── Utilities ──────────────────────────────────────────────
tqdm>=4.65.0
